+++ TA-ms-teams-alert-action/bin/ta_ms_teams_alert_action/modalert_ms_teams_publish_to_channel_helper.py
@@ -196,6 +196,14 @@ def process_event(helper, *args, **kwargs):
     # data facts
     data_json_facts = '\"facts\": [\n'
 
+    # Fields ordering in the message publication, defaults to alphabetical ordering
+    alert_ms_teams_fields_order = helper.get_param("alert_ms_teams_fields_order")
+    alert_ms_teams_fields_order_by_alpha = True
+    helper.log_debug("alert_ms_teams_fields_order={}".format(alert_ms_teams_fields_order))
+    if alert_ms_teams_fields_order in "order_by_list":
+        alert_ms_teams_fields_order_by_alpha = False
+    helper.log_debug("alert_ms_teams_fields_order_by_alpha={}".format(alert_ms_teams_fields_order_by_alpha))
+
     # Iterate over the results to extract key values from the field list provided in input
     alert_ms_teams_fields_list = helper.get_param("alert_ms_teams_fields_list")
     helper.log_debug("alert_ms_teams_fields_list={}".format(alert_ms_teams_fields_list))
@@ -222,14 +230,19 @@ def process_event(helper, *args, **kwargs):
                     else:
                         keys_dict_str = '"' + checkstr(key) + '": ' + '"' + checkstr(value) + '"'
                     keys_count += 1
+            helper.log_debug("keys_dict_str={}".format(keys_dict_str))
 
             # Convert to a proper dict
             keys_dict_str = '{' + keys_dict_str + '}'
-            helper.log_debug("keys_dict_str={}".format(keys_dict_str))
             keys_dict = json.loads(keys_dict_str)
+            helper.log_debug("Before processing ordering, keys_dict_str={}".format(keys_dict_str))
 
-            # Order alphabetically by key
-            keys_ordered = OrderedDict(sorted(keys_dict.items(), key=lambda t: t[0]))
+            # Ordering fields
+            if alert_ms_teams_fields_order_by_alpha:
+                keys_ordered = OrderedDict(sorted(keys_dict.items(), key=lambda t: t[0]))
+            else:
+                keys_ordered = OrderedDict(sorted(keys_dict.items(), key=lambda pair: alert_ms_teams_fields_list.index(pair[0])))
+            helper.log_debug("After processing ordering as instructed, keys_ordered={}".format(keys_ordered))
 
             for key, value in keys_ordered.items():
                 if key in alert_ms_teams_fields_list: